backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
django-filter==2.3.0
Pillow==10.4.0